add uuid validator

+++ internal/repositories/organizations.go
@@ -5,7 +5,9 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
+	"github.com/donskova1ex/application_aggregator/internal"
 	"github.com/donskova1ex/application_aggregator/internal/domain"
+	"github.com/donskova1ex/application_aggregator/tools"
 	"github.com/google/uuid"
 	"time"
 )
@@ -35,13 +37,17 @@ func (repo *PostgresRepository) CreateOrganization(ctx context.Context, organiza
 }
 
 func (repo *PostgresRepository) GetOrganizationByUUID(ctx context.Context, uuid string) (*domain.Organization, error) {
+	if !tools.ValidUUID(uuid) {
+		return nil, fmt.Errorf("invalid uuid: %w", internal.UUIDValidationFailed)
+	}
+
 	query := `SELECT uuid, name FROM organizations WHERE uuid = $1`
 
 	organization := &domain.Organization{}
 
 	err := repo.db.GetContext(ctx, organization, query, uuid)
 	if errors.Is(err, sql.ErrNoRows) {
-		return nil, nil
+		return nil, fmt.Errorf("organization with uuid not found: %w", internal.ErrRecordNotFound)
 	}
 	if err != nil {
 		return nil, fmt.Errorf("error getting organization by uuid: %w", err)
@@ -51,16 +57,34 @@ func (repo *PostgresRepository) GetOrganizationByUUID(ctx context.Context, uuid 
 }
 
 func (repo *PostgresRepository) DeleteOrganizationByUUID(ctx context.Context, uuid string) error {
+	if !tools.ValidUUID(uuid) {
+		return fmt.Errorf("invalid uuid: %w", internal.UUIDValidationFailed)
+	}
+
 	querty := `DELETE FROM organizations WHERE uuid = $1`
 
-	_, err := repo.db.ExecContext(ctx, querty, uuid)
+	result, err := repo.db.ExecContext(ctx, querty, uuid)
+
 	if err != nil {
-		return fmt.Errorf("there is no organization with this uuid: %w", err)
+		return fmt.Errorf("error deleting organization with this uuid: %w", err)
+	}
+
+	rowsAffected, err := result.RowsAffected()
+	if err != nil {
+		return fmt.Errorf("error checking rows affected: %w", err)
+	}
+
+	if rowsAffected == 0 {
+		return fmt.Errorf("organization with uuid not found: %w", internal.ErrRecordNotFound)
 	}
 
 	return nil
 }
 func (repo *PostgresRepository) UpdateOrganization(ctx context.Context, uuid string, organization *domain.Organization) (*domain.Organization, error) {
+	if !tools.ValidUUID(uuid) {
+		return nil, fmt.Errorf("invalid uuid: %w", internal.UUIDValidationFailed)
+	}
+
 	query := `UPDATE organizations SET name = $1, updated_at = $2 WHERE uuid = $3`
 
 	updateTime := time.Now()
